Better Python displayhook
Get rid of return value display for _display_hook_ methods and rename
interface to _displayhook_ since it makes more sense and we might as
well while we are breaking backwards compatibility.  This also means we
I can get rid of that silly _ReprWrapper class.  I also made it so that
the _ variable only gets set for non-None values, like it does with the
default displayhook, which is the reason I came back to the displayhook
stuff in the first place.