Better dummy file in dummy directory
An actual icon directory needs to exist for it to be a valid theme, and
we need something to put in it.  Earlier, I had a .directory file that
was automatically created by Dolphin, but that seemed awkward and like
something that I might end up ignoring by accident, so I changed it.